--- src/mcfunction/functions/Functions.ts
@@ -1,8 +1,0 @@
-export function GetCommand(Text: string) : string {
-    var Index = Text.indexOf(" ");
-
-    if (Index <= 0)
-        return "";
-
-    return Text.substring(0, Index);
-}
