+++ src/mcfunction/main.ts
@@ -1,12 +1,16 @@
 import * as vscode from 'vscode';
 import * as McfunctionCompletionItems from './completion_item';
+import * as Selector from "./selectors/completion_item"
 
 export function activate(context: vscode.ExtensionContext) {
 	// triggered whenever a ' ' is being typed
-	const CompletionProvider = vscode.languages.registerCompletionItemProvider(
+	context.subscriptions.push(vscode.languages.registerCompletionItemProvider(
 		'bc-minecraft-mcfunction', 
 		new McfunctionCompletionItems.McFunctionCompletion(), 
-		" ", "\n" );
+		" ", "\n" ));
 
-	context.subscriptions.push(CompletionProvider);
+	context.subscriptions.push(vscode.languages.registerCompletionItemProvider(
+			'bc-minecraft-mcfunction', 
+			new Selector.SelectorCompletion(), 
+			"@", "a", "e", "r", "p", "s", "[", ","));
 }
