+++ tsconfig.json
@@ -9,6 +9,7 @@
 		"noImplicitAny": true,
 		"removeComments": true,
 		"preserveConstEnums": true,
+		"resolveJsonModule": true
 	},
 	"exclude": [
 		"node_modules",
